refactor(Prompts): move dialog components to dialogs/ subdirectory
