+++ client/src/components/Prompts/dialogs/VariableDialog.tsx
@@ -3,7 +3,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog';
 import type { TPromptGroup } from 'librechat-data-provider';
 import { OGDialog, OGDialogTitle, OGDialogContent } from '@librechat/client';
 import { detectVariables } from '~/utils';
-import VariableForm from './VariableForm';
+import VariableForm from '../forms/VariableForm';
 
 interface VariableDialogProps extends Omit<DialogPrimitive.DialogProps, 'onOpenChange'> {
   onClose: () => void;
