+++ client/src/components/Prompts/dialogs/index.ts
@@ -1,0 +1,4 @@
+export { default as PreviewPrompt } from './PreviewPrompt';
+export { default as DeleteVersion } from './DeleteVersion';
+export { default as VariableDialog } from './VariableDialog';
+export { default as SharePrompt } from './SharePrompt';
